Fix update script so that it builds
The update script had to be adapted as well as the BigInt library so
that MSVC now builds everything. Additionally, we also have to fix
common XML-ization module in tests.

Previously, a couple of issues in BigInt and common XML-ization
prevented the compilation. Namely, we compared signed and unsigned
values. The update script was too sparse on messages and we could not
figure out what went wrong.

+++ test/common_xmlization.cpp
@@ -532,7 +532,7 @@ std::string CanonicalizeXml(
   XML_Status status = XML_Parse(
     parser,
     xml.data(),
-    xml.size(),
+    static_cast<int>(xml.size()),
     true
   );
